+++ prinz_mlflow/src/test/scala/pl/touk/nussknacker/prinz/MLFContainerUnitTest.scala
@@ -4,13 +4,13 @@ import java.io.File
 import java.time.Duration
 import com.dimafeng.testcontainers.WaitingForService
 import org.testcontainers.containers.wait.strategy.LogMessageWaitStrategy
-import pl.touk.nussknacker.prinz.MLFUnitIntegrationTest.{MLFLOW_SERVER_SERVICE_NAME,
+import pl.touk.nussknacker.prinz.MLFContainerUnitTest.{MLFLOW_SERVER_SERVICE_NAME,
   MODEL_1_SERVED_READY_REGEX, MODEL_2_SERVED_READY_REGEX, MODEL_3_SERVED_READY_REGEX,
   TIMEOUT_MINUTES, readEnvWithName}
 import pl.touk.nussknacker.prinz.container.ContainerUnitTest
 import pl.touk.nussknacker.prinz.container.ContainerUnitTest.EnvMap
 
-abstract class MLFUnitIntegrationTest extends ContainerUnitTest {
+abstract class MLFContainerUnitTest extends ContainerUnitTest {
 
   override def dockerComposeFile: File = new File("../dev-environment/docker-compose-mlflow.yaml")
 
@@ -30,7 +30,7 @@ abstract class MLFUnitIntegrationTest extends ContainerUnitTest {
   )
 }
 
-object MLFUnitIntegrationTest {
+object MLFContainerUnitTest {
 
   private val TIMEOUT_MINUTES = 5
